fixed an error in load_doc_2_word
